+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ### Fixed
 
+- Another project_id scope issue in Otu Filter
 - Update DwcOccurence index endpoint
 
 ## [0.37.0] - 2023-12-14